refactor(benchmarks): split core/benchmark.py into data.benchmark spec + benchmarks/ harness
Split the 1728-line core/benchmark.py two ways:
- SPEC (a dataset IS a benchmark) -> src/langres/data/benchmark.py (import-light:
  pydantic + typing + PairMetrics/pairs_from_clusters). Carries Benchmark protocol,
  PairTrack, gold_pairs_from_clusters, complete_partition, BlindCostError,
  DEFAULT_PAIR_GRID, _Resolvable/RecordT.
- HARNESS -> new src/langres/benchmarks/ package: runner.py (run_method(s), tables,
  tracks, resolver-bcubed, tune_threshold, cost helpers) + judge_eval.py (JudgePairEval,
  evaluate(_judge_on_candidates), BudgetedModuleRunner, truncation). __init__ exports
  nothing (import-light like autoresearch).

core/benchmark.py is now a TEMPORARY re-export shim (pragma: no cover) so old imports
resolve. data/* import the spec from data.benchmark DIRECTLY (never via the core shim)
to avoid a core<->data cycle; er_benchmarks imports the harness resolver-bcubed
function-locally to keep data -> benchmarks off the module top.

Repointed all src consumers; eval.py facade lazy map -> data.benchmark (spec) +
benchmarks.judge_eval (evaluate). Tangle baseline: core.benchmark -> benchmarks.runner
(same {.,methods} 2-cycle; count 18->18, largest SCC 9->9). refactor json adds the
langres.benchmarks prefix.
